# services/twilight/internal/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "twilight",
    "version": "1",
    "description": "Run campaigns on the fleet, read its inventory and alerts. Authenticate with a login session (cookie, plus the twilight_csrf cookie's value in X-CSRF-Token on every request that changes something), an API token (Authorization: Bearer), or a client certificate. Every error is {\"error\": {\"code\", \"message\", \"details\"}}. Lists are pages of {\"items\", \"next_cursor\"}."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearer": []
    },
    {
      "session": []
    },
    {
      "certificate": []
    }
  ],
  "paths": {
...
    "/api/v1/devices/{device}/lifecycle": {
      "post": {
        "operationId": "setDeviceLifecycle",
        "summary": "Change a device's lifecycle",
        "x-required-role": "admin",
        "description": "Retiring or revoking a device blocks every installation of it, now and later. active lifts the block, and each installation keeps its own lifecycle. The change is produced to dusk.node-state, which nightfall enforces, and then written to the inventory.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Device"
          },
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceLifecycleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The device's lifecycle after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceLifecycle"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/selectors/validate": {
      "post": {
        "operationId": "validateSelector",
        "summary": "Check a selector",
        "x-required-role": "viewer",
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Whether it parses, where it fails, how many nodes it matches and a sample of them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectorValidation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/campaigns": {
      "get": {
        "operationId": "listCampaigns",
        "summary": "List campaigns",
        "x-required-role": "viewer",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Only campaigns in these statuses, comma-separated or repeated.",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of campaigns, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "operationId": "createCampaign",
        "summary": "Create a campaign",
        "x-required-role": "operator",
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignDefinition"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new draft campaign.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "The campaign's path.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/campaigns/{id}": {
      "get": {
        "operationId": "getCampaign",
        "summary": "A campaign",
        "x-required-role": "viewer",
        "parameters": [
          {
            "$ref": "#/components/parameters/Campaign"
          }
        ],
        "responses": {
          "200": {
            "description": "The campaign.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "put": {
        "operationId": "updateCampaign",
        "summary": "Edit a draft campaign",
        "x-required-role": "operator",
        "description": "Only a draft can be edited. 409 with code version_conflict when the campaign changed since the version given was read.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Campaign"
          },
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The campaign after the edit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
    "/api/v1/campaigns/{id}/start": {
      "post": {
        "operationId": "startCampaign",
        "summary": "Start a draft campaign",
        "x-required-role": "operator",
        "description": "409 with code overlap when an ensure campaign of the same kind is running on some of the same nodes; details give the count and the campaigns. Set policy.allow_overlap to start anyway.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Campaign"
          },
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The campaign after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/campaigns/{id}/pause": {
      "post": {
        "operationId": "pauseCampaign",
        "summary": "Pause a running campaign",
        "x-required-role": "operator",
        "description": "Stops new dispatches; processes in flight finish.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Campaign"
          },
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The campaign after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/campaigns/{id}/resume": {
      "post": {
        "operationId": "resumeCampaign",
        "summary": "Resume a paused campaign",
        "x-required-role": "operator",
        "description": "409 with code gate_override_required after a health gate paused it, unless override_gate is true and a reason is given.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Campaign"
          },
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The campaign after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/campaigns/{id}/abort": {
      "post": {
        "operationId": "abortCampaign",
        "summary": "Abort a campaign",
        "x-required-role": "operator",
        "description": "Stops new dispatches and cancels the rows not yet dispatched.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Campaign"
          },
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The campaign after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceLifecycle": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{32}$"
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "active",
              "retired",
              "revoked"
            ]
          },
          "reason": {
            "type": "string"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time"
          },
          "actor": {
            "type": "string",
            "description": "Who made the change."
          }
        },
        "required": [
          "device_id",
          "lifecycle",
          "reason",
          "changed_at",
          "actor"
        ],
        "description": "The last lifecycle change made to a device as a whole."
      },
...
      "DeviceLifecycleRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "lifecycle": {
            "type": "string",
            "enum": [
              "active",
              "retired",
              "revoked"
            ]
          },
          "reason": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "lifecycle",
          "reason"
        ]
      },
...
    },
...
    "parameters": {
      "Device": {
        "name": "device",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-f]{32}$"
        }
      },
...
      "CSRF": {
        "name": "X-CSRF-Token",
        "in": "header",
        "description": "Required with a session cookie: the value of the twilight_csrf cookie.",
        "schema": {
          "type": "string"
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token from twilight token create."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "twilight_session",
        "description": "A login session from /api/v1/auth/login."
      },
      "certificate": {
        "type": "mutualTLS",
        "description": "A client certificate from the internal CA naming one urn:dusk:principal, mapped to a role by principals."
      }
    }
  }
}